+++ src/lib/andromeda/filesystem/filedata/MemoryAllocator.hpp
@@ -12,7 +12,12 @@ namespace Andromeda {
 namespace Filesystem {
 namespace Filedata {
 
-// TODO !! comments (only does PAGE_SIZE chunks (64K on windows), no caching), yes thread safe
+/**
+ * A raw, non-caching memory allocator that allocates pages directly from the OS, bypassing the C library
+ * NOTE memory is allocated only at page size granularity.  Use get_usage() to determine the actual memory size of an allocation.
+ * In DEBUG builds, verifies all calls to free() for validity.
+ * THREAD SAFE (INTERNAL LOCKS)
+ */
 class MemoryAllocator
 {
 public:
@@ -23,18 +28,24 @@ public:
     virtual ~MemoryAllocator();
 #endif // DEBUG
 
+    /** Allocate the given number of bytes and return a pointer */
     void* alloc(size_t bytes);
 
+    /**
+     * Free a memory allocation returned by alloc().
+     * @param bytes the number of bytes allocated (must match)
+     */
     void free(void* const ptr, size_t bytes);
 
     /** 
-     * Returns the actual memory used by an allocation
+     * Calculates the actual memory used by an allocation
      * Allocation sizes are always rounded by page granularity
      */
     size_t get_usage(const size_t bytes);
 
 private:
 
+    /** Updates and prints allocator statistics (debug) */
     void stats(const std::string& fname, const size_t bytes, bool alloc);
 
 #if DEBUG
@@ -43,17 +54,16 @@ private:
     AllocMap mAllocMap;
 #endif // DEBUG
 
-    /** total number of bytes allocated */
-    uint64_t mTotal { 0 };
-    /** total number of allocs */
+    /** total number of bytes allocated (debug) */
+    size_t mTotal { 0 };
+    /** total number of allocs (debug) */
     uint64_t mAllocs { 0 };
-    /** total number of frees */
+    /** total number of frees (debug) */
     uint64_t mFrees { 0 };
     /** stat-counter mutex */
     std::mutex mMutex;
 
     typedef std::lock_guard<decltype(mMutex)> LockGuard;
-
     Debug mDebug;
 };
